requirements.txt
django-oscar>=2.0,<2.1
wagtail>=2.5,<2.6
django>=2.2,<2.3

pytest-cov>=2.3.1
pytest-django>=3.0.0
pytest-pythonpath>=0.7
pytest>=3.0.3
